Improve dark mode toggle, add color transitions to body
