+++ app/components/dark-mode-toggle.tsx
@@ -1,44 +1,11 @@
 'use client';
 
+import { useDarkMode } from '@/hooks/use-dark-mode';
 import { MoonIcon, SunIcon } from '@heroicons/react/24/solid';
-import { useEffect, useState } from 'react';
 import { CircleIcon } from './icons/circle-icon';
 
 export function DarkModeToggle() {
-    const [isDarkMode, setIsDarkMode] = useState<boolean>();
-    const [prefersDarkMode, setPrefersDarkMode] = useState<boolean>();
-
-    useEffect(() => {
-        const darkMediaQuery = window.matchMedia(
-            '(prefers-color-scheme: dark)'
-        );
-
-        const updateDarkMode = () => {
-            const isDark = document.documentElement.classList.contains('dark');
-            setIsDarkMode(isDark);
-            setPrefersDarkMode(darkMediaQuery.matches);
-        };
-
-        updateDarkMode();
-        darkMediaQuery.addEventListener('change', updateDarkMode);
-        window.addEventListener('storage', updateDarkMode);
-
-        return () => {
-            darkMediaQuery.removeEventListener('change', updateDarkMode);
-            window.removeEventListener('storage', updateDarkMode);
-        };
-    }, []);
-
-    const toggleDarkMode = () => {
-        const isDark = document.documentElement.classList.toggle('dark');
-        setIsDarkMode(isDark);
-
-        if (isDark === prefersDarkMode) {
-            localStorage.removeItem('isDarkMode');
-        } else {
-            localStorage.setItem('isDarkMode', JSON.stringify(isDark));
-        }
-    };
+    const { isDarkMode, toggleDarkMode } = useDarkMode();
 
     if (isDarkMode === undefined) {
         return (
